decoupled exporter/tracing setup functions
